pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepmine"
version = "0.1.0"
description = "Deep Earth Exploration Pipeline for Microbial INhibitor Extraction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "DEEPMINE Contributors"},
]
keywords = [
    "metagenomics",
    "antibiotic-discovery",
    "biosynthetic-gene-clusters",
    "bioinformatics",
    "machine-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "biopython>=1.83",
    "snakemake>=8.0",
    "pandas>=2.1",
    "numpy>=1.26",
    "torch>=2.1",
    "scikit-learn>=1.4",
    "rdkit>=2023.9.1",
    "gecco>=0.9",
    "requests>=2.31",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
gpu = [
    "torch>=2.1",
]
docs = [
    "sphinx>=7.2",
    "sphinx-rtd-theme>=2.0",
]

[project.scripts]
deepmine = "deepmine.cli:main"

[project.urls]
Homepage = "https://github.com/deepmine-project/deepmine"
Repository = "https://github.com/deepmine-project/deepmine"
Issues = "https://github.com/deepmine-project/deepmine/issues"
Documentation = "https://deepmine.readthedocs.io"

[tool.setuptools.packages.find]
include = ["deepmine*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["deepmine"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external resources",
    "gpu: marks tests requiring GPU",
]
